NumberGuessGame.py
This small game allows the user to guess a number between two ends and then drills down to that number using a binary search algorithm. It takes a guess and everytime it gets a response back from the user it partitions on the data set it has using binary search algorithm to find the number that the user initially thought of.